requirements.txt
shellfoundry
cloudshell-automation-api